requirements.txt
Django==1.5.1
distribute==0.6.34
dj-database-url==0.2.1
gunicorn==0.17.4
wsgiref==0.1.2
django-tastypie
